pyproject.toml
[project]
name = "my-qrz-maps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "folium",
  "mercantile",
  "pillow",
  "requests",
]
